+++ ReactNativeStudy/Lisa/Week4/map/src/components/SvgIcons.tsx
@@ -1,7 +1,7 @@
 import {Path, Svg, SvgProps} from 'react-native-svg';
 
-import {KeyOfPalette, theme} from '../../styles/theme';
-import {svgIconPath} from '../../constants/svgIconPath';
+import {svgIconPath} from '../constants/svgIconPath';
+import {KeyOfPalette, theme} from '../styles/theme';
 
 interface SvgIconsProps extends SvgProps {
   iconVariant: keyof typeof svgIconPath;
