+++ ReactNativeStudy/Lisa/Week4/map/src/screens/Favorite.tsx
