+++ ReactNativeStudy/Lisa/Week4/map/src/screens/Home.tsx
@@ -1,4 +1,4 @@
-import Map from '../../components/Map';
+import Map from '../components/Map';
 
 const Home = () => {
   return <Map />;
